Several enhancements to the Word Emitter (see bugs Bug 431616 and 433725):  * Support for word "fields" (formulas)  * avoid superfluos nested table All of this is controlled by the following user properties: PROP_NAME_LAYOUTMODE = "WordEmitter.LayoutMode" PROP_NAME_FIELDFUNCTION = "WordEmitter.FieldFunction" PROP_NAME_NESTED_TABLE = "WordEmitter.NestedTable"
By adding WordEmitter.LayoutMode = "NONE" to a list, the emitter avoids creating word tables for the list.

By adding WordEmitter.FieldFunction = "any non-empty string" to a label, the emitter uses the label's content as a field function.

WordEmitter.NestedTable is a boolean property and usually works automatic: It avoids a nested 1x1 table for a "foreign" element (dynamic text)
if this element is in a cell of a table or grid.
Note that the border properties of the element are ignored then.
By setting the value to true, it forces the old behaviour.

Furthermore, the DOC emitter now supports "show header on first".