PDF emitter VerticalTab user property (#2020)
Added support for a new UserProperty `PdfEmitter.VerticalTab` to the PDF emitter.
The property can be set for cells and text items/labels/data and is used to go down at least a given *fixed distance* measured from the top of the page body, independent from the height of already was written further above on the page.